Created RegisterForm and added it to page
